fixing travis
