+++ benchmarks/CMakeLists.txt
@@ -16,6 +16,6 @@ function(make_folly_benchmark NAME FILE)
     ${GLOG_LIBRARY})
 endfunction()
 
-make_folly_benchmark(streamthroughput StreamThroughput.cpp)
-make_folly_benchmark(baselines_tcp BaselinesTcp.cpp)
-make_folly_benchmark(baselines_async_socket BaselinesAsyncSocket.cpp)
+make_folly_benchmark(streamthroughput StreamThroughput.cpp Benchmarks.cpp)
+make_folly_benchmark(baselines_tcp BaselinesTcp.cpp Benchmarks.cpp)
+make_folly_benchmark(baselines_async_socket BaselinesAsyncSocket.cpp Benchmarks.cpp)
